refactor: split Map.cs (1,450→1,203 lines) — extract save serialization to partial class
Extracted all map save/serialization methods from Map.cs into
Map.Save.cs using a C# partial class. Zero behavioral change.

Map.cs (1,203 lines): core map class, helpers (BinaryDataHeader,
  MapField, MapVisibility), constructors, terrain/cell math,
  terrain projection, resize, IReadOnlyFileSystem, Dispose
Map.Save.cs (273 lines): Save(), SaveBinaryData(), SavePreview(),
  GetTerrainColorPair(), GetCellSpaceBounds()

The save logic was a clearly separable concern — map binary format
serialization, preview PNG generation, and cell space bounds
calculation are only used during map saving.
